Use yum priorities plugin rather than injecting Epoch
This means we don't need to modify many spec files to handle the
`Epoch` tag being optionally present.

http://lists.rpm.org/pipermail/rpm-ecosystem/Week-of-Mon-20150713/000125.html

The downside is we need to do more manipulation of the mock
configuration which could be fragile over time, and we need to pull in
the priorities plugin (which will be loaded on the host too).