Support building using specfiles from upstream
Obviously it's very convenient if one has specfiles in
an upstream git repo.

This is a step towards a more generic framework...although I'm
hesitant to go down the plugins route.  Let's keep things more
hardcoded for now.

Obviously the first test case of this was using rpmdistro-gitoverlay
to test build itself.