Move the srpm logic into resolve
Now the workdir structure looks like:

 - src: Git mirror
 - snapshot: Directory of SRPMs plus snapshot.json build config
 - build{,0,1}: Binary RPMs

When we added the `prep-cmd` to the overlay schema, we started
fetching sources during build which is wrong.  We might as well bite
the bullet here and just use SRPMs as the canonical representation
here.

This is preparatory work for being able to consume pure dist-git
repositories, and pulling in `fedpkg/pyrpkg` more closely.