Add new `self-buildrequires` key
In Fedora right now, the `redhat-rpm-config` package depends on
`go-srpm-macros`, which I'm trying to build.  Since rrc is in the
buildroot, this means `go-srpm-macros` self-BRs (sadly).  We can work
around this by assuming that we have a previous build - we just need
to teach rpmdistro-gitoverlay to copy it into the root.

This isn't the default because this model shouldn't be encouraged.