build: Fix losing previous successful builds
In changing things to retain partial builds, we started walking
over the build list after it was complete, and deleting components
that weren't marked successful.

But builds we didn't need to do weren't marked as successful!  Fix
this by retaining a set of builds we need to do, and don't try to
postprocess builds we didn't make.

This means we won't delete the cache metadata for previously
successful builds.