resolve: Add --override-giturl
When wiring rdgo to Jenkins and receving git push notifications
(e.g. github webhook), it's convenient to pass them on to resolve via
this command line flag, without having to maintain an external mapping
of which components contain which git URLs.

This allows overriding the branch (e.g. to use the `auto` branch for Homu),
or optionally fetching from an alternative location, and to fetch an
exact commit.

The first case as mentioned is for watching one upstream branch
(`master`, `auto`, `try` etc.).  This second case is intended for
Jenkins PR testing, where Jenkins watches for `refs/pr/*`, and we
inherit the fetch from Jenkins.

Closes: #30
Approved by: cgwalters