Support 'with', 'without' keys
These ultimately filter down to `rpmbuild --with|out`.  The rationale
for this is that for some components I actually want to make
use of this, even though it's not traditional in Fedora/CentOS.

For example, rpm-ostree can be built with/without bundled libhif, and
for rdgo I want to not bundle.

Closes: #32
Approved by: cgwalters