Add a 'srpmroot' key, used to inject SRPM root dependencies
I'm trying to build a newer golang from Fedora dist-git for CentOS
Atomic Host Continuous, but hitting an issue where it's relying on the
implicit `redhat-rpm-config` -> `go-srpm-macros` in Fedora, which
doesn't exist in CentOS.

We can work around that by specially denoting RPMs which contain
macros, and ensuring they end up in the buildroot.  This is cleaner
than defining our own `foo-rpm-config` package.