resolve: Add an override-version tag
So before this I had a relatively hard line in that the
RPM version-release tags should match `git describe`.  Anyone
who wants to ensure lower versions win should be using the
`priority=` key on repos.

Which works well with yum, but it turns out we don't implement
it in libhif, and hence rpm-ostree doesn't support it.

This is currently causing `skopeo-1.9` from base RHEL to win over our
CI-generated RPMs in CAHC.

We should teach libhif about priorities, but in the meantime this hack
gets us by.