Rework to run in Fedora, not CentOS
https://pagure.io/fedora-atomic-host-continuous/issue/16

Fedora started using richdeps in the `BuildRequires`, which finally
brought down our hacky tower to do Fedora RPM builds from a CentOS 7
host.

Complicating all of this is that mock in Fedora is python3.  So
we'd have to maintain compatibility with both 2/3 and that sucks.
Let's do a hard 3 port for now.